copy changes to Apply page Relates #100
