+++ src/pages/apply.js
@@ -1,5 +1,4 @@
-import React, { Component, Fragment } from "react"
-import styled from "styled-components"
+import React, { Component } from "react"
 import * as r from "ramda"
 
 import { APPLICATION_CAROUSEL, DOWN_CURSOR } from "../constants"
@@ -7,17 +6,18 @@ import { APPLICATION_CAROUSEL, DOWN_CURSOR } from "../constants"
 import Layout from "../components/Layout"
 import { Panel, NextPanel, ExtendedPanel } from "../components/Panel"
 import ApplicationsStatus from "../components/ApplicationsStatus"
-import Step from "../components/Step"
+
 import Table from "../components/Table"
 import Carousel from "../components/Carousel"
 import { Cursor } from "../components/Cursor"
 import HeadingWithBody from "../components/HeadingWithBody"
-import { BigUnderline, _SubHeading } from "../components/Text"
+import { BigUnderline, _Heading } from "../components/Text"
 import DoubleLine from "../components/DoubleLine"
 import FlickityCarousel from "../components/FlickityCarousel"
 import BackgroundImg from "../components/BackgroundImg"
 import InnerGridContainer from "../components/InnerGridContainer"
 import { CollapsableQuestion, QuestionWrapper } from "../components/Question"
+import ApplicationSteps from "../components/StudentApplicationSteps"
 
 import FAC10 from "../assets/photos/FAC10.jpg"
 import FAC13 from "../assets/photos/FAC13.jpg"
@@ -25,91 +25,20 @@ import FAC14 from "../assets/photos/FAC14.jpg"
 import FACN3 from "../assets/photos/FACN3.png"
 import FACG3 from "../assets/photos/FACG3.png"
 
-const ApplicationSteps = () => (
-  <Fragment>
-    <Step step="01" colour="yellow" title="Submit your expression of interest">
-      Please submit an expression of interest to get invited to events,
-      workshops and meetups for prospective students. We open expressions of
-      interest six months before a cohort begins.
-      <br/><br/>
-      <a href="">Submit expression of Interest</a>
-    </Step>
-    <Step
-      step="02"
-      colour="blue"
-      title="Start work on our course prerequisites"
-    >
-      <span className="b">Before we can consider your application</span>, 
-      you will need to complete our course prerequisites. We encourage
-      applicants to take as much time as they need to work through the material
-      and to support each other during the preparation process, both online and
-      through our meetups.
-    </Step>
-    <Step>
-      <div>
-        <_SubHeading>Create a GitHub account</_SubHeading>
-        You will use this account to join Gitter and to create your one-page
-        site on GitHub Pages.
-      </div>
-      <div>
-        <_SubHeading>FreeCodeCamp</_SubHeading>
-        Earn a minimum of 200 points on freeCodeCamp. We recommend beginning
-        with the following sections:
-        <h3 className="underline">Responsive Web Design Certification</h3>
-        <ul>
-          <_ListItem>Basic HTML and HTML5</_ListItem>
-          <_ListItem>Basic CSS</_ListItem>
-          <_ListItem>Responsive Web Design Principles</_ListItem>
-          <_ListItem>CSS Flexbox</_ListItem>
-        </ul>
-        <h3 className="underline">
-          Javascript Algorithms and Data Structures Certification
-        </h3>
-        <ul>
-          <_ListItem>Basic JavaScript</_ListItem>
-          <_ListItem>ES6</_ListItem>
-          <_ListItem>Basic Data Structures</_ListItem>
-          <_ListItem>Basic Algorithm Scripting</_ListItem>
-        </ul>
-      </div>
-    </Step>
-    <Step step="03" colour="red" title="Submit an application">
-      <p>
-        Use [prereq-check] to ensure you are prepared to submit your
-        application. Plan to spend about an hour completing the application
-        form.
-      </p>
-      <p>
-        If you submit your application before the deadline, we encourage you to
-        keep learning and tidying up your website. We will evaluate your
-        progress as of the date the window closes.
-      </p>
-    </Step>
-    <Step step="04" colour="green" title="The conversational interview" last={true}>
-      We invite a minimum of 24 people to interview (in person or remotely), and{" "}
-      <span className="b">offer places to sixteen people per cohort</span>.
-      You’ll be interviewed by a panel of Founders and Coders staff and alumni.
-    </Step>
-  </Fragment>
-)
-const _ListItem = styled.li`
-  list-style-type: circle;
-  list-style-position: inside;
-`
 
 const CarouselImg = ({ src }) => (
   <BackgroundImg src={src} height="calc(10rem + 25vw)" width="100%" gradient="radial-gradient( rgba(141,140,355,0.25), rgba(4,2,171,0.25) 120%)"/>
 )
 
 const carouselImages = [
-  { caption: "London Cohort #10", src: FAC10 },
-  { caption: "Nazareth Cohort #3", src: FACN3 },
+  { caption: "London Cohort #14", src: FAC14 },
   { caption: "Gaza Cohort #4", src: FACG3 },
   { caption: "London Cohort #13", src: FAC13 },
-  { caption: "London Cohort #14", src: FAC14 },
+  { caption: "Nazareth Cohort #3", src: FACN3 },
+  { caption: "London Cohort #10", src: FAC10 },
 ]
 
-const initialCarouselIdx = 2
+const initialCarouselIdx = 0
 
 class ApplyPage extends Component {
   state = {
@@ -205,7 +134,7 @@ class ApplyPage extends Component {
                 </CollapsableQuestion>
               </QuestionWrapper>
               <div className="font-5 fw3">
-                * Applications close at midnight GMT on the final day of the window.
+                Applications close at midnight GMT on the final day of the window.
               </div>
             </HeadingWithBody>
           </InnerGridContainer>
@@ -233,6 +162,7 @@ class ApplyPage extends Component {
           </section>
 
           <InnerGridContainer className="mb7 pb5">
+            <_Heading className="ml6 mb4">Join our developer community</_Heading>
             <HeadingWithBody title={this.state.carouselCaption}>
               <FlickityCarousel
                 options={{
@@ -240,8 +170,6 @@ class ApplyPage extends Component {
                   prevNextButtons: false,
                   initialIndex: initialCarouselIdx,
                 }}
-                // width="953pxx
-                // height="568px"
                 changeCaption={caption =>
                   this.setState({ carouselCaption: caption })
                 }
